Improve dashboard UX with routing, clearer workflows, and component extraction.
